moved source inside package and updated refs
