+++ README.md
@@ -23,7 +23,7 @@ GPT Engineer is made to be easy to adapt, extend, and make your agent learn how 
 **Run**:
 - Create a new empty folder with a `main_prompt` file (or copy the example folder `cp example -r my-new-project`)
 - Fill in the `main_prompt` in your new folder
-- run `python main.py my-new-project`
+- run `poetry run gpt_engineer my-new-project`
 
 **Results**:
 - Check the generated files in my-new-project/workspace_clarified
